feat: check if the user exist or not

+++ login/register.js
@@ -1,9 +1,19 @@
-
+import getUsers from "./users.js"
 
 // Adding a new user
 function addUser(newUser) {
-  const storedUsers = localStorage.getItem('users');
-  let usersArray = JSON.parse(storedUsers) || [];
-  usersArray.push(newUser);
-  localStorage.setItem('users', JSON.stringify(usersArray));
+  let exist = false;
+
+  getUsers.forEach(element => {
+    if (user.email === newUser.email) {
+      exist = true;
+      return;
+    }
+  });
+
+  if (!exist) {
+    usersArray.push(newUser);
+    localStorage.setItem('users', JSON.stringify(usersArray));
+  }
+
 }
